requirements.txt
aiohttp>=3.9.0
PyYAML>=6.0.1
numpy>=1.24.0
pandas>=2.0.0
MetaTrader5>=5.0.45
